MA-56: Put edge button outside of node
