+++ Views/NodeView.axaml
@@ -1,9 +1,8 @@
-<controls:NodeUserControl
+<UserControl
 	xmlns="https://github.com/avaloniaui"
 	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
 	xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
 	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-	xmlns:controls="clr-namespace:mystery_app.Controls"
 	xmlns:vm="clr-namespace:mystery_app.ViewModels"
 	xmlns:views="clr-namespace:mystery_app.Views"
 	xmlns:constants ="clr-namespace:mystery_app.Constants"
@@ -16,15 +15,12 @@
 		<vm:NodeViewModel />
 	</Design.DataContext>
 
-	<Grid RowDefinitions="Auto,*,Auto,Auto,Auto">
-		<Panel Grid.Row="0" Tag="EdgeButton" Width="10" Height="10" Margin="5" Background="Red"
-			PointerPressed="SelectNodeEdge"
-			PointerReleased="ReleaseNodeEdge"/>
-		<views:DragDropImageView Grid.Row="1" DataContext="{Binding Image}"/>
-		<TextBox Grid.Row="2" Text="{Binding Name}" DockPanel.Dock="Bottom"/>
-		<TextBox Grid.Row="3" Text="{Binding Desc}" DockPanel.Dock="Bottom"/>
-		<Expander Grid.Row="4" Header="Notes" VerticalAlignment="Top">
+	<Grid RowDefinitions="*,Auto,Auto,Auto">
+		<views:DragDropImageView Grid.Row="0" DataContext="{Binding Image}"/>
+		<TextBox Grid.Row="1" Text="{Binding Name}" DockPanel.Dock="Bottom"/>
+		<TextBox Grid.Row="2" Text="{Binding Desc}" DockPanel.Dock="Bottom"/>
+		<Expander Grid.Row="3" Header="Notes" VerticalAlignment="Top">
 			<TextBox Text="{Binding Notes}"/>
 		</Expander>
 	</Grid>
-</controls:NodeUserControl>
+</UserControl>
