+++ Views/WorkspaceView.axaml
@@ -78,7 +78,7 @@
 				<ItemsControl.ItemTemplate>
 					<DataTemplate>
 						<views:InteractiveView DataContext="{Binding}">
-							<views:NodeView DataContext="{Binding}" />
+							<views:NodeView DataContext="{Binding}" Grid.Row="1"/>
 						</views:InteractiveView>
 					</DataTemplate>
 				</ItemsControl.ItemTemplate>
